+++ src/main/java/ru/shadrindmitry/diploma/restaurantvoting/to/PositionInMenuTo.java
@@ -8,9 +8,9 @@ import lombok.Value;
 public class PositionInMenuTo extends BaseTo {
 
     String description;
-    Double price;
+    int price;
 
-    public PositionInMenuTo(Integer id, String description, Double price) {
+    public PositionInMenuTo(Integer id, String description, int price) {
         super(id);
         this.description = description;
         this.price = price;
